Tests/LibWeb/Text/expected/css/transition-on-element-keeps-pseudo-element-font.txt

FAIL: ::after width was changed by the transition
